// src/app/services/api-backend/model/role.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RolesUsuario } from './rolesUsuario';


export interface Role { 
    colRol?: number;
    nombreRol?: string | null;
    estado?: string | null;
    usuarioIng?: string | null;
    fechaIng?: string;
    usuarioAct?: string | null;
    fechaAct?: string | null;
    rolesUsuarios?: Array<RolesUsuario> | null;
}